요구사항이 이미지 확대인데 카드확대로 하여서 수정하였고, api호출부분 분리
